using mergo instead

+++ service/controller/chart/resource/release/desired_test.go
@@ -137,7 +137,7 @@ func Test_DesiredState(t *testing.T) {
 				ValuesMD5Checksum: "4845bfb2cf922d7527886ac13599ea3b",
 				Values: map[string]interface{}{
 					"provider": "azure",
-					"replicas": 2,
+					"replicas": float64(2),
 				},
 				Version: "0.1.2",
 			},
@@ -226,7 +226,7 @@ func Test_DesiredState(t *testing.T) {
 				Status:            helmclient.StatusDeployed,
 				ValuesMD5Checksum: "2187a8fce91c3765a74d462062af7526",
 				Values: map[string]interface{}{
-					"secretnumber":   2,
+					"secretnumber":   float64(2),
 					"secretpassword": "admin",
 				},
 				Version: "0.1.2",
@@ -297,8 +297,8 @@ func Test_DesiredState(t *testing.T) {
 				Status:            helmclient.StatusDeployed,
 				ValuesMD5Checksum: "40f61add2413651115ef94f85d8335cf",
 				Values: map[string]interface{}{
-					"replicas":     2,
-					"secretnumber": 2,
+					"replicas":     float64(2),
+					"secretnumber": float64(2),
 					"username":     "admin",
 				},
 				Version: "0.1.2",
